Tabs for pre-processing, ROI selection and DIC solver are finished
It is possible now to start from a set of raw images, generate a set of images that are cropped in a more interesting field of view. Then it is possible to generate the FEM with the GUI and define all the settings of the muDIC solver interactively. We can run then the solver with a simple click.
Next part is the post-processing with a GUI to get the actual displacement. The user would also be in a position to keep the pixel values of displacement.
It will be possible to select the different variables that can be visualized.
It seems valuable to add two different complementary tabs. One will be dedicated to the preparation of test campaigns depending on the camera lenses, distance between the sample and the camera. It will provide a possibility to visualize the filed of view. The choice of the speckle to apply on the sample will be guided be some calculations accounting for the optical properties and the magnitude of the quantities of interest that are expected.
Another tab will be added to prepare dialogue with the Open Source FEM code developed by EDF: Code_Aster. It will provide the possibility to generate boundary conditions and a Code_Aster script to carry out a simulation of the test.